fix: KEEP-444 keep http-request/step.ts as a pure step boundary
CI build failed: the Workflow DevKit pulled lib/safe-fetch.ts into
the workflow-function bundle, which forbids Node modules
(undici / node:net / node:async_hooks). The DevKit treats a step
file as a boundary -- only `"use step"` exports are stubbed into
the workflow-function bundle; everything else exported from the
same file gets analyzed in the workflow bundle along with its
transitive imports. Exporting `httpRequest` (which calls safeFetch)
from step.ts is what tripped the analyzer.

Moves all the worker logic (httpRequest, the timeout/failOnError
resolvers, validators, parse helpers, types, safeFetch import) into
`lib/workflow/nodes/http-request/perform.ts`. `step.ts` is reduced
to a thin `"use step"` wrapper and type re-exports. Tests import
from `./perform`.

No behavior change. `pnpm build` now succeeds locally (was failing
in CI on the same workflow-bundling error this PR introduced).
